Big sweep to make SGCT compile in MacOS again  - Updated GLFW version to to missing CarbonSound header on newer MacOS versions  - Fixed errors in #ifdef'd code  - Removed warnings reported by Clang

+++ include/sgct/SGCTTrackingDevice.h
@@ -55,15 +55,15 @@ public:
     int getNumberOfAxes() const;
 
     /// \returns a digital value from array
-    bool getButton(size_t index) const;
+    bool getButton(int index) const;
     
     /// \returns a digital value from array
-    bool getButtonPrevious(size_t index) const;
+    bool getButtonPrevious(int index) const;
 
     /// \returns an analog value from array
-    double getAnalog(size_t index) const;
+    double getAnalog(int index) const;
     /// \returns an analog value from array
-    double getAnalogPrevious(size_t index) const;
+    double getAnalogPrevious(int index) const;
     
     bool isEnabled() const;
     bool hasSensor() const;
